+++ coroc/coroc_volatile_test.go
@@ -8,7 +8,7 @@ import (
 	"github.com/stealthrocket/coroutine"
 )
 
-func assertSerializable[R, S any](t *testing.T, g coroutine.Generator[R, S]) coroutine.Generator[R, S] {
+func assertSerializable[R, S any](t *testing.T, g coroutine.Coroutine[R, S]) coroutine.Coroutine[R, S] {
 	// Not serializable unless the "durable" build tag is present.
 	return g
 }
