+++ coroc/go.mod
@@ -2,10 +2,12 @@ module github.com/stealthrocket/coroutine/coroc
 
 go 1.21.0
 
-require golang.org/x/tools v0.13.0
+require (
+	github.com/stealthrocket/coroutine v0.0.0-20230906012022-7474cda88ddc
+	golang.org/x/tools v0.13.0
+)
 
 require (
-	github.com/stealthrocket/coroutine v0.0.0-20230906002520-a57a3bcc65df // indirect
 	golang.org/x/mod v0.12.0 // indirect
 	golang.org/x/sys v0.12.0 // indirect
 )
